+++ apps/main/src/init.ts
@@ -19,10 +19,6 @@ import { initializeSentry } from "./sentry"
 import { router } from "./tipc"
 import { createMainWindow, getMainWindow } from "./window"
 
-const appFolder = {
-  prod: "Follow",
-  dev: "Follow (dev)",
-}
 if (process.argv.length === 3 && process.argv[2].startsWith("follow-dev:")) {
   process.env.NODE_ENV = "development"
 }
@@ -32,7 +28,8 @@ const isDev = process.env.NODE_ENV === "development"
  * Mandatory and fast initializers for the app
  */
 export function initializeAppStage0() {
-  app.setPath("appData", path.join(app.getPath("appData"), isDev ? appFolder.dev : appFolder.prod))
+  if (isDev) app.setPath("appData", path.join(app.getPath("appData"), "Follow (dev)"))
+  initializeSentry()
 }
 export const initializeAppStage1 = () => {
   if (process.defaultApp) {
@@ -44,8 +41,6 @@ export const initializeAppStage1 = () => {
   } else {
     app.setAsDefaultProtocolClient(APP_PROTOCOL)
   }
-
-  initializeSentry()
 
   registerIpcMain(router)
 
@@ -150,12 +145,20 @@ const registerPushNotifications = async () => {
   updateNotificationsToken()
 
   const instance = new PushReceiver({
-    debug: isDev,
-    firebase: env.VITE_FIREBASE_CONFIG,
+    debug: true,
+    firebase: JSON.parse(env.VITE_FIREBASE_CONFIG),
     persistentIds: persistentIds || [],
-    credentials,
+    credentials: credentials || null,
+    bundleId: "is.follow",
+    chromeId: "is.follow",
   })
-  logger.info(`PushReceiver initialized with token ${credentials?.fcm?.token}`)
+  logger.info(
+    `PushReceiver initialized with credentials ${JSON.stringify(credentials)} and firebase config ${env.VITE_FIREBASE_CONFIG}`,
+  )
+
+  instance.onReady(() => {
+    logger.info("PushReceiver ready")
+  })
 
   instance.onCredentialsChanged(({ newCredentials }) => {
     logger.info(`PushReceiver credentials changed to ${newCredentials?.fcm?.token}`)
@@ -195,5 +198,11 @@ const registerPushNotifications = async () => {
     store.set(persistentIdsKey, instance.persistentIds)
   })
 
-  await instance.connect()
+  try {
+    await instance.connect()
+  } catch (error) {
+    logger.error(`PushReceiver error: ${error instanceof Error ? error.stack : error}`)
+  }
+
+  logger.info("PushReceiver connected")
 }