+++ apps/renderer/src/modules/discover/user-form.tsx
@@ -27,7 +27,7 @@ export function DiscoverUser() {
   return (
     <>
       {data?.follow.routes && (
-        <div className="w-[512px]">
+        <div className="w-[540px]">
           <DiscoverFeedForm
             routePrefix="follow"
             route={data.follow.routes[Object.keys(data.follow.routes)[0]]}